requirements.txt
numpy
mediapipe
Pillow
opencv-python
torch >= 1.0.0
torchvision